# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Felt API Doc",
        "description": "Felt API Documentation",
        "contact": {
            "email": "rejohnbd@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer/vehicle-documents": {
            "get": {
                "tags": [
                    "vehicle-documents"
                ],
                "summary": "Returns all customer documents",
                "description": "return customer documents",
                "operationId": "get-vehicle-documents",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "vehicle-documents"
                ],
                "summary": "Store Vehicle Document",
                "description": "Store Vehicle Document.",
                "operationId": "vehicle-documents",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vehicle_paper_id[]",
                        "in": "path",
                        "description": "Vehicle Paper Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "expire_date[]",
                        "in": "path",
                        "description": "Expire Paper Id",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "document_image[]",
                        "in": "path",
                        "description": "Document Image ",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}